+++ crates/integrations/connector-integration/src/connectors/razorpay/test.rs
@@ -116,6 +116,7 @@ mod tests {
                     l2_l3_data: None,
 
                     merchant_request_id: None,
+                    sender_payment_instrument_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Razorpay {
                     api_key: "dummy_api_key".to_string().into(),
@@ -308,7 +309,7 @@ mod tests {
                     minor_amount_capturable: None,
                     amount: None,
                     l2_l3_data: None,
-
+                    sender_payment_instrument_id: None,
                     merchant_request_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Razorpay {
@@ -453,7 +454,7 @@ mod tests {
                     minor_amount_capturable: None,
                     amount: None,
                     l2_l3_data: None,
-
+                    sender_payment_instrument_id: None,
                     merchant_request_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Razorpay {
@@ -620,8 +621,8 @@ mod tests {
                     minor_amount_capturable: None,
                     amount: None,
                     l2_l3_data: None,
-
                     merchant_request_id: None,
+                    sender_payment_instrument_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Razorpay {
                     api_key: "dummy_api_key".to_string().into(),
@@ -985,7 +986,7 @@ mod tests {
                 order_details: None,
                 minor_amount_capturable: None,
                 l2_l3_data: None,
-
+                sender_payment_instrument_id: None,
                 merchant_request_id: None,
             },
             connector_config: ConnectorSpecificConfig::Razorpay {
@@ -1186,8 +1187,8 @@ mod tests {
                 recurring_mandate_payment_data: None,
                 order_details: None,
                 l2_l3_data: None,
-
                 merchant_request_id: None,
+                sender_payment_instrument_id: None,
             },
             connector_config: ConnectorSpecificConfig::Razorpay {
                 api_key: "dummy_api_key".to_string().into(),
@@ -1389,7 +1390,7 @@ mod tests {
                     recurring_mandate_payment_data: None,
                     order_details: None,
                     l2_l3_data: None,
-
+                    sender_payment_instrument_id: None,
                     merchant_request_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Razorpay {
@@ -1510,8 +1511,8 @@ mod tests {
                     recurring_mandate_payment_data: None,
                     order_details: None,
                     l2_l3_data: None,
-
                     merchant_request_id: None,
+                    sender_payment_instrument_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Razorpay {
                     api_key: "dummy_api_key".to_string().into(),
@@ -1634,7 +1635,7 @@ mod tests {
                     recurring_mandate_payment_data: None,
                     order_details: None,
                     l2_l3_data: None,
-
+                    sender_payment_instrument_id: None,
                     merchant_request_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Razorpay {
@@ -1800,8 +1801,8 @@ mod tests {
                 order_details: None,
                 minor_amount_capturable: None,
                 l2_l3_data: None,
-
                 merchant_request_id: None,
+                sender_payment_instrument_id: None,
             },
             connector_config: ConnectorSpecificConfig::Razorpay {
                 api_key: "dummy_api_key".to_string().into(),
@@ -1933,7 +1934,7 @@ mod tests {
                 recurring_mandate_payment_data: None,
                 order_details: None,
                 l2_l3_data: None,
-
+                sender_payment_instrument_id: None,
                 merchant_request_id: None,
             },
             connector_config: ConnectorSpecificConfig::Razorpay {
@@ -2055,7 +2056,7 @@ mod tests {
                 recurring_mandate_payment_data: None,
                 order_details: None,
                 l2_l3_data: None,
-
+                sender_payment_instrument_id: None,
                 merchant_request_id: None,
             },
             connector_config: ConnectorSpecificConfig::Razorpay {