+++ CHANGELOG.md
@@ -19,6 +19,69 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - - -
 
+## 2026.05.19.1
+
+### Features
+
+- **connector:**
+  - Implement Reverse (VoidPostCapture) for fiservemea ([#1301](https://github.com/juspay/connector-service/pull/1301)) ([`bdb37e0`](https://github.com/juspay/connector-service/commit/bdb37e04f56154acedae77cd4c0ce051e6fa5a2a))
+  - Implement Reverse (VoidPostCapture) for adyen ([#1277](https://github.com/juspay/connector-service/pull/1277)) ([`4656d5c`](https://github.com/juspay/connector-service/commit/4656d5c22c0ceada5004a8cb366e0577330961f8))
+  - Implement Reverse (VoidPostCapture) for cybersource ([#1278](https://github.com/juspay/connector-service/pull/1278)) ([`0d48651`](https://github.com/juspay/connector-service/commit/0d48651caca2d64514fc31add3dcc208493e9e7f))
+
+### Refactors
+
+- **proto:** Group wallet payment methods by flow type ([#1356](https://github.com/juspay/connector-service/pull/1356)) ([`df6b261`](https://github.com/juspay/connector-service/commit/df6b2611c8227a2e43f2d8c0c2b9f99c0c567151))
+
+### Testing
+
+- **nuvei:** Harden integration test overrides ([#1299](https://github.com/juspay/connector-service/pull/1299)) ([`3286746`](https://github.com/juspay/connector-service/commit/32867461eaabdbfa58810404d00c7c064eb9ec7c))
+
+**Full Changelog:** [`2026.05.19.0...2026.05.19.1`](https://github.com/juspay/connector-service/compare/2026.05.19.0...2026.05.19.1)
+
+- - -
+
+## 2026.05.19.0
+
+### Features
+
+- **connector:**
+  - Implement Reverse (VoidPostCapture) for jpmorgan ([#1281](https://github.com/juspay/connector-service/pull/1281)) ([`c33f182`](https://github.com/juspay/connector-service/commit/c33f182db4c95c4bbce96b9d5caf88da46894d41))
+  - Implement Reverse (VoidPostCapture) for authipay ([#1298](https://github.com/juspay/connector-service/pull/1298)) ([`af473c4`](https://github.com/juspay/connector-service/commit/af473c4c8758ae6a07d88b6a0d8f83f8c44cb71b))
+
+### Bug Fixes
+
+- **connector:** [Truelayer] Populate sender_payment_instrument_id from psync response ([#1362](https://github.com/juspay/connector-service/pull/1362)) ([`79cfb69`](https://github.com/juspay/connector-service/commit/79cfb69eaa559081f9ded1dc98ce5201bd0e874d))
+
+**Full Changelog:** [`2026.05.18.1...2026.05.19.0`](https://github.com/juspay/connector-service/compare/2026.05.18.1...2026.05.19.0)
+
+- - -
+
+## 2026.05.18.1
+
+### Features
+
+- **shadow:** Add x-merchant-id header to outgoing shadow mode requests ([#1310](https://github.com/juspay/connector-service/pull/1310)) ([`cce18ed`](https://github.com/juspay/connector-service/commit/cce18ed0742a90f3312861139d74933b247d0403))
+
+### Bug Fixes
+
+- **airwallex:** Fix parity for psync and createorder 5 field-mapping divergences ([#1213](https://github.com/juspay/connector-service/pull/1213)) ([`4cadc99`](https://github.com/juspay/connector-service/commit/4cadc99e96185c6b6f3c5c92bbdb756d1b1d3e9a))
+- **connector:** [twoc_twop_paco] rename creq to raw_creq and add raw_connector_response ([#1355](https://github.com/juspay/connector-service/pull/1355)) ([`a40bcec`](https://github.com/juspay/connector-service/commit/a40bceccd6b7a707a5385cfb27b78062a350662c))
+- Added Euler-UCS mass enabled gateways in params patch ([#1316](https://github.com/juspay/connector-service/pull/1316)) ([`42b26c2`](https://github.com/juspay/connector-service/commit/42b26c29de685736a69cbbf3ee246ea164499d22))
+
+### Testing
+
+- **bankofamerica:** Harden integration test overrides ([#1318](https://github.com/juspay/connector-service/pull/1318)) ([`305c44c`](https://github.com/juspay/connector-service/commit/305c44c1d7bbd33ff4c1852203525f9c3b811daa))
+- **billwerk:** Harden integration test overrides ([#1280](https://github.com/juspay/connector-service/pull/1280)) ([`f889be3`](https://github.com/juspay/connector-service/commit/f889be331167f1fb6a84a92e6fbea489bc87ebe0))
+- **cybersource:** Harden integration test overrides ([#1284](https://github.com/juspay/connector-service/pull/1284)) ([`776d4e9`](https://github.com/juspay/connector-service/commit/776d4e9bc121871329fbc207e733a3d5a667cc31))
+- **globalpay:** Harden integration test overrides ([#1290](https://github.com/juspay/connector-service/pull/1290)) ([`1864dc9`](https://github.com/juspay/connector-service/commit/1864dc9396da12f28001c915594f5716a09053cb))
+- **stax:** Harden integration test overrides ([#1312](https://github.com/juspay/connector-service/pull/1312)) ([`688cb05`](https://github.com/juspay/connector-service/commit/688cb05a08eb946f7b653597613a7f681b860181))
+- **trustpay:** Harden integration test overrides ([#1313](https://github.com/juspay/connector-service/pull/1313)) ([`31bf23d`](https://github.com/juspay/connector-service/commit/31bf23d25b2e4b7589a0933c09be32dd46cdd56f))
+- **worldpayvantiv:** Harden integration test overrides ([#1315](https://github.com/juspay/connector-service/pull/1315)) ([`1777410`](https://github.com/juspay/connector-service/commit/1777410c74785775c10f9291a689344c484af6f9))
+
+**Full Changelog:** [`2026.05.18.0...2026.05.18.1`](https://github.com/juspay/connector-service/compare/2026.05.18.0...2026.05.18.1)
+
+- - -
+
 ## 2026.05.18.0
 
 ### Features